Add files via upload
Code for a coffee machine that works with every value of coin. And has three menu items (Espresso, Latte, Cappuccino).
It will refund the difference if extra coin is inserted and also, will refund all if the inserted coin is not enough for the transaction. 
It will report the value of the ingredients if the personnel asks...